fix: Improve fastboot userspace reboot & sparse limits
Switch SparseMaxDownloadSize to a long and default it to uint.MaxValue, and clamp computed download sizes against both the host-side limit and the protocol (uint32) max. Adjust CLI and tests to use the new type and to apply the correct min/max. Improve reboot-to-userspace behavior: shorten the reconnect pause, print a waiting message, use EnsureUserspace for fastboot-target reboots, and add clearer error messages plus a final IsUserspace check after reconnect. Also bump FirmwareKit.Sparse to 1.0.1 and remove an AVB package reference.

+++ FirmwareKit.Comm.Fastboot.Tests/FastbootInfoTests.cs
@@ -199,7 +199,7 @@ namespace FirmwareKit.Comm.Fastboot.Tests
         [Fact]
         public void GetMaxDownloadSize_ClampsToSparseLimit()
         {
-            int originalLimit = FastbootDriver.SparseMaxDownloadSize;
+            long originalLimit = FastbootDriver.SparseMaxDownloadSize;
             try
             {
                 FastbootDriver.SparseMaxDownloadSize = 1024 * 1024 * 1024;
